Move `InMemoryFeedStore` to it's own file
