+++ FeedFetchApp/FeedFetchAppTests/FeedAcceptanceTests.swift
@@ -95,56 +95,3 @@ final class FeedAcceptanceTests: XCTestCase {
     }
     
 }
-
-
-// MARK:- InMemoryFeedStore
-
-class InMemoryFeedStore {
-    private(set) var feedCache: CachedFeed?
-    private var feedImageDataCache: [URL: Data] = [:]
-
-    private init(feedCache: CachedFeed? = nil) {
-        self.feedCache = feedCache
-    }
-}
-
-extension InMemoryFeedStore: FeedStore {
-    func deleteCachedFeed(completion: @escaping FeedStore.DeletionCompletion) {
-        feedCache = nil
-        completion(.success(()))
-    }
-
-    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping FeedStore.InsertionCompletion) {
-        feedCache = CachedFeed(feed: feed, timestamp: timestamp)
-        completion(.success(()))
-    }
-
-    func retrieveCachedFeed(completion: @escaping RetrievalCompletion) {
-        completion(.success(feedCache))
-    }
-}
-
-extension InMemoryFeedStore: FeedImageDataStore {
-    func retrieveImageData(for url: URL, completion: @escaping (RetrieveResult) -> Void) {
-        completion(.success(feedImageDataCache[url]))
-    }
-    
-    func insertImageData(_ data: Data, for url: URL, completion: @escaping (InsertResult) -> Void) {
-        feedImageDataCache[url] = data
-        completion(.success(()))
-    }
-}
-
-extension InMemoryFeedStore {
-    static var empty: InMemoryFeedStore {
-        InMemoryFeedStore()
-    }
-
-    static var withExpiredFeedCache: InMemoryFeedStore {
-        InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date.distantPast))
-    }
-
-    static var withNonExpiredFeedCache: InMemoryFeedStore {
-        InMemoryFeedStore(feedCache: CachedFeed(feed: [], timestamp: Date()))
-    }
-}
